requirements.txt
# Core Python
python-dotenv>=1.0.0
Flask>=2.0.1

# LLM Providers
google-generativeai>=0.3.0
groq>=0.3.0

# Math & Symbolic Computation
sympy>=1.12

# Production
# gunicorn>=20.1.0  # Uncomment for production

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Utilities
pydantic>=2.6.0
typing-extensions>=4.9.0
gunicorn>=20.1.0
waitress>=2.1.2